requirements.txt
aiosqlite==0.20.0
aiogram==3.12.0
python-decouple==3.8